client/requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
idna==3.10
ocrmac==1.0.0
pillow==11.0.0
pyobjc-core==10.3.1
pyobjc-framework-cocoa==10.3.1
pyobjc-framework-coreml==10.3.1
pyobjc-framework-quartz==10.3.1
pyobjc-framework-vision==10.3.1
requests==2.32.3
urllib3==2.2.3